+++ lib/main_db/core/daos/base/bank_card/bank_card_items_dao.dart
@@ -72,4 +72,8 @@ class BankCardItemsDao extends DatabaseAccessor<MainStore>
         .getSingleOrNull();
     return row?.read(bankCardItems.routingNumber);
   }
+
+  Future<void> upsertBankCardItem(BankCardItemsCompanion companion) {
+    return into(bankCardItems).insertOnConflictUpdate(companion);
+  }
 }