+++ lib/main_db/core/daos/base/crypto_wallet/crypto_wallet_items_dao.dart
@@ -64,4 +64,8 @@ class CryptoWalletItemsDao extends DatabaseAccessor<MainStore>
         .getSingleOrNull();
     return row?.read(cryptoWalletItems.xprv);
   }
+
+  Future<void> upsertCryptoWalletItem(CryptoWalletItemsCompanion companion) {
+    return into(cryptoWalletItems).insertOnConflictUpdate(companion);
+  }
 }