+++ lib/main_db/core/daos/base/loyalty_card/loyalty_card_items_dao.dart
@@ -66,4 +66,8 @@ class LoyaltyCardItemsDao extends DatabaseAccessor<MainStore>
         .getSingleOrNull();
     return row?.read(loyaltyCardItems.password);
   }
+
+  Future<void> upsertLoyaltyCardItem(LoyaltyCardItemsCompanion companion) {
+    return into(loyaltyCardItems).insertOnConflictUpdate(companion);
+  }
 }