+++ lib/main_db/core/daos/base/password/password_items_dao.dart
@@ -47,4 +47,8 @@ class PasswordItemsDao extends DatabaseAccessor<MainStore>
         .getSingleOrNull();
     return row?.read(passwordItems.password);
   }
+
+  Future<void> upsertPasswordItem(PasswordItemsCompanion companion) {
+    return into(passwordItems).insertOnConflictUpdate(companion);
+  }
 }