+++ lib/main_db/core/daos/base/ssh_key/ssh_key_items_dao.dart
@@ -47,4 +47,8 @@ class SshKeyItemsDao extends DatabaseAccessor<MainStore>
         .getSingleOrNull();
     return row?.read(sshKeyItems.privateKey);
   }
+
+  Future<void> upsertSshKeyItem(SshKeyItemsCompanion companion) {
+    return into(sshKeyItems).insertOnConflictUpdate(companion);
+  }
 }